Made migrations for Order model

+++ app/core/models.py
@@ -230,7 +230,7 @@ class Order(models.Model):
     title = models.CharField(max_length=255)
     description = models.TextField(null=True, blank=True)
     provider = models.ForeignKey(Provider, on_delete=models.CASCADE, related_name="records")
-    deal_value = models.IntegerField(blank=True)
+    deal_value = models.DecimalField(max_digits=12, decimal_places=2, blank=True)
     currency = models.SmallIntegerField(choices=CURRENCY_TYPE_CHOICES, default=UAH)
     file = models.FileField(upload_to=str(config('RECORDS_URL')))
     office = models.ForeignKey(Office, on_delete=models.CASCADE, related_name="records")
